Revert "Merge code of Group 107"
